rename app folder
