+++ busy_beaver/apps/upcoming_events/upcoming_events.py
@@ -1,6 +1,6 @@
 import time
 
-from busy_beaver.apps.events.cards import UpcomingEventList
+from busy_beaver.apps.upcoming_events.cards import UpcomingEventList
 from busy_beaver.common.wrappers.meetup import EventDetails
 from busy_beaver.models import Event, UpcomingEventsConfiguration
