+++ busy_beaver/blueprints.py
@@ -1,6 +1,6 @@
 from busy_beaver.apps.debug.api.healthcheck import healthcheck_bp  # noqa
-from busy_beaver.apps.events.blueprint import events_bp  # noqa
 from busy_beaver.apps.github_integration.blueprint import github_bp  # noqa
 from busy_beaver.apps.retweeter.blueprint import twitter_bp  # noqa
 from busy_beaver.apps.slack_integration.api import slack_bp  # noqa
+from busy_beaver.apps.upcoming_events.blueprint import events_bp  # noqa
 from busy_beaver.apps.web.blueprint import web_bp  # noqa
